authModule/requirements.txt
alembic==1.17.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
Deprecated==1.3.1
Flask==3.1.2
Flask-Limiter==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==5.21.0
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
rich==14.2.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
wrapt==2.0.1
